+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.5.0
 - Adapt Data Contract Specification in version 0.9.1
 - Use `models` section for `diff`/`breaking`
+- Add `model` command
 
 # 0.4.0
 - Basic implementation of `test` command for Soda Core